+++ pyodide-build/pyodide_build/out_of_tree/build.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 
 from .. import common, pypabuild, pywasmcross
 
@@ -10,6 +11,9 @@ def run(exports, args):
     cxxflags += f" {os.environ.get('CXXFLAGS', '')}"
     ldflags = common.get_make_flag("SIDE_MODULE_LDFLAGS")
     ldflags += f" {os.environ.get('LDFLAGS', '')}"
+
+    curdir = Path.cwd()
+    (curdir / "dist").mkdir(exist_ok=True)
     build_env_ctx = pywasmcross.get_build_env(
         env=os.environ.copy(),
         pkgname="",